move BlockCreationTime into its own module
