+++ src/Chainweb/BlockHeader/Genesis.hs
@@ -39,6 +39,7 @@ import Pact.Types.Hash (Hash)
 
 -- internal modules
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeight
