+++ src/Chainweb/Cut/Test.hs
@@ -79,6 +79,7 @@ import qualified Test.QuickCheck.Monadic as T
 
 -- internal modules
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeight
