+++ src/Chainweb/Logging/Amberdata.hs
@@ -63,6 +63,7 @@ import System.LogLevel
 
 -- internal modules
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeight
