+++ src/Chainweb/Pact/Backend/ForkingBench.hs
@@ -82,6 +82,7 @@ import Pact.Types.Util hiding (unwrap)
 
 -- chainweb imports
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeader.Genesis
