+++ src/Chainweb/Pact/PactService.hs
@@ -104,6 +104,7 @@ import Pact.Types.Term (Term(..))
 ------------------------------------------------------------------------------
 -- internal modules
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeader.Genesis (genesisBlockHeader, genesisBlockPayload)
