+++ src/Chainweb/Pact/Service/BlockValidation.hs
@@ -29,6 +29,7 @@ import Data.Vector (Vector)
 import Pact.Types.Command
 import Pact.Types.Hash
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeight
