+++ src/Chainweb/Pact/Service/Types.hs
@@ -35,6 +35,7 @@ import Pact.Types.Hash
 
 -- internal chainweb modules
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeight
