+++ src/Chainweb/Pact/Types.hs
@@ -133,6 +133,7 @@ import Pact.Types.Term (PactId(..), Ref)
 
 -- internal chainweb modules
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeight
