+++ src/Chainweb/RestAPI/Orphans.hs
@@ -46,6 +46,7 @@ import Servant.API
 
 -- internal modules
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeaderDB
