+++ src/Chainweb/WebPactExecutionService/Types.hs
@@ -11,6 +11,7 @@ import Data.Vector (Vector)
 import Pact.Types.Command
 import Pact.Types.Hash
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeight
