+++ test/Chainweb/Test/Pact/PactInProcApi.hs
@@ -46,6 +46,7 @@ import Pact.Parse
 import Pact.Types.ChainMeta
 import Pact.Types.Command
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHeader
 import Chainweb.BlockHeader.Genesis
 import Chainweb.ChainId
