+++ test/Chainweb/Test/Pact/TTL.hs
@@ -34,6 +34,7 @@ import Pact.Types.ChainMeta
 
 -- chainweb imports
 
+import Chainweb.BlockCreationTime
 import Chainweb.BlockHash
 import Chainweb.BlockHeader
 import Chainweb.BlockHeight
